Represent lines as u32 and columns as u16
With this change we represent source location lines as u32 and columns
as u16. Doing so is less wasteful than using machine words for both and
it is what both our GSYM and DWARF resolvers use internally. On top of
that doing so provides us with a few padding bytes inside the
symbolize::Sym type that we can use for other purposes, should the need
arise.

+++ include/blazesym.h
@@ -266,10 +266,11 @@ typedef struct blaze_sym {
    */
   const char *path;
   /**
-   * The line number on which the symbol was to be found in the source code.
+   * The line number on which the symbol is located in the source
+   * code.
    */
-  size_t line;
-  size_t column;
+  uint32_t line;
+  uint16_t column;
 } blaze_sym;
 
 /**
